Extract shared payments route helpers
